Implement Treaty Handler
Also:
- finished implementTreaty()
- moved foodRange, treatyValid and implementTreaty functions to treaties.go
